+++ src/analysis/mod.rs
@@ -4,6 +4,10 @@
 //!
 //! This module implements various analysis passes on the IR.
 
+mod domtree;
+mod preds;
 mod trg;
 
+pub use self::domtree::*;
+pub use self::preds::*;
 pub use self::trg::*;