+++ src/pass/vtpp.rs
@@ -2,9 +2,7 @@
 
 //! Var to Phi Promotion
 
-use crate::ir::prelude::*;
-use crate::opt::prelude::*;
-use crate::pass::gcse::PredecessorTable;
+use crate::{analysis::PredecessorTable, ir::prelude::*, opt::prelude::*};
 use std::collections::{HashMap, HashSet};
 
 /// Var to Phi Promotion